[FEAT] #29 SSE 관련 파일 분리
